requirements.txt
Django>=3.2, <3.3
python-dotenv
djangorestframework
requests
django-jsonfield